fix(warnings): eliminate all MSVC C4189/C4101/C4100/C4505/C4244/C4838 warnings
src/hal/rf_hardware.h|cpp: constructor int->int instead of uint8_t->uint8_t
  (eliminates C4244 narrowing across all test files at once)

tests/test_nbap_edch.cpp:  inline bool ok into assert(); [[maybe_unused]] r/r1/r2
tests/test_nr_phy.cpp:     inline sfnMsn into assert expression
tests/test_oms.cpp:        [[maybe_unused]] id2; drop id3 assignment; (void)a in loop
tests/test_pcap_writer.cpp: [[maybe_unused]] readBE16/readBE32/readLE16;
  pull f.read() / gh read / probe read out of asserts (fix C4100/C4101)
tests/test_pdcp.cpp:       inline ct pointer into memcmp call
tests/test_pm_export.cpp:  remove unused clearCounters() function
tests/test_umts_mac.cpp:   [[maybe_unused]] ue1/ue2/ue3/ue4
tests/test_umts_phy.cpp:   inline rscp into assert expression
tests/test_umts_rlc.cpp:   [[maybe_unused]] r1; inline gotSdu into assert
tests/test_umts_rrc.cpp:   [[maybe_unused]] r2; fix 384000->384 kbps (C4838)
tests/test_umts_stack.cpp: (now covered by constructor fix)

All 43/43 CTests pass.

+++ src/hal/rf_hardware.cpp
@@ -6,9 +6,9 @@
 
 namespace rbs::hal {
 
-RFHardware::RFHardware(uint8_t txAntennas, uint8_t rxAntennas)
-    : txAntennas_(txAntennas)
-    , rxAntennas_(rxAntennas)
+RFHardware::RFHardware(int txAntennas, int rxAntennas)
+    : txAntennas_(static_cast<uint8_t>(txAntennas))
+    , rxAntennas_(static_cast<uint8_t>(rxAntennas))
     , activeTxMask_((1u << txAntennas) - 1u)
 {}
